% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATTS-package.R
\docType{package}
\name{BATTS-package}
\alias{BATTS}
\alias{BATTS-package}
\title{Boosting and Bayesian Additive Trees for Two-Sample Comparison}
\description{
Estimate the square root of a density ratio using boosting or generalized
Bayesian additive trees. See \code{\link[=boots]{boots()}}, \code{\link[=batts]{batts()}}, and
\code{\link[=eval_balance_weight]{eval_balance_weight()}} for the public interface.
}
\references{
Awaya, N., Xu, Y., and Ma, L. (2025). Two-sample comparison
through additive tree models for density ratios.
\url{https://arxiv.org/abs/2508.03059}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nawaya040/BATTS}
  \item Report bugs at \url{https://github.com/nawaya040/BATTS/issues}
}

}
\author{
\strong{Maintainer}: Naoki Awaya \email{nawaya@waseda.jp}

Authors:
\itemize{
  \item Yuliang Xu
  \item Li Ma
}

}
